+++ examples/cloud_storage/variables.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2021 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-terraform {
-  required_providers {
-    google = {
-      source  = "hashicorp/google"
-      version = "~> 4.1"
-    }
-  }
-  required_version = ">= 0.13"
+variable "project_id" {
+  type        = string
+  description = "The project ID to manage the Pub/Sub resources"
 }